LayoutTests/imported/w3c/web-platform-tests/content-security-policy/worker-src/dedicated-worker-src-child-fallback-blocked.sub-expected.txt


FAIL Same-origin dedicated worker allowed by worker-src 'self'. assert_equals: expected "worker-src" but got "child-src"
